Move the file HistogramConfigPanelComponent.tsx to folder HistogramSettingsPanelComponent
