+++ src/components/Histogram/HistogramSettingsPanelComponent/HistogramConfigPanelComponent.tsx
